requirements.txt
streamlit
google-generativeai
supabase
psycopg2-binary
pandas